Measures put in place that keep outsiders from taking surveys and keep insiders from taking surveys they shouldn't be taking.
Assign Surveys page fixed to list only currently assigned surveys, my surveys, and public surveys.

Survey selection changed to a submit from a link, so it is more indicative of an action being taken rather than a link being followed; and also so parameters are passed through POST instead of GET, so it is less inviting for rapscallions to try and take inappropriate surveys through URL manipulation.

"(not logged in)" put in place of the user's name while under an anonymous survey-taking session.

git-svn-id: https://svn.csc.ncsu.edu:8443/svn/research-expertiza/trunk@90 392d7b7b-3f31-0410-9dc3-c95a9635ea79