Fix spontaneous logout / permission issues
In newer versions of Rails, when a form is submitted without an
authenticity token, it clears the session rather than raising an
exception.

Some requests (such as the broken impersonate autocomplete) were doing
AJAX requests that triggered this and logged us out, when we normally
wouldn't have seen the exception.