Copy features from branches A and B as samples
Could not copy the step definitions, as they were implemented in Watir
instead of capybara. These will have to be re-implemented.

Some of the features are duplicates, as two independent teams worked on
this without collaboration, and their work was combined.

These samples should be removed as real features that cover the same
functionality are written.