The following objects have been refactored as part of the CSC517 refactoring project (E406):
/config/environments/development.rb
/app/controllers/password_retrieval_controller.rb
/test/functional/password_retrieval_controller_test.rb
/app/views/password_retrieval/forgotten.html.erb

The motivation for the refactoring was that the password retrieval
controller only had a single method which contained the implementation
details of numerous functionality such as validating the email, resetting
the password, and sending an email.

As it was currently written, it would have been difficult to extend or
modify the functionality such as changing how email is actually sent or
altering the validation of the email address to include more validation steps.

Refactoring has allowed for separation of concerns, enhanced
maintainability, and made the code more extensible.