This commit refactored survey_controller.rb In general, the assign method was extensive, long, and hard to follow. As such, we extracted some methods and renamed a few variables for clarity. It is important to note that as the refactoring was taking place, the team noticed that this method was not working as designed.
Spefically, a previous refactoring effort to the database failed to correct the
implementation details of the assign method. Specifically, the assign method had SQL which referenced a column named type_id. However, this
column was removed with the database refactoring effort as can be seen
in /db/migrate/107_merge_questionnaire_and_type.rb

Now, survey_controller.rb works as designed. Thus, this effort encompassed
both a bug fix and a refactoring effort.