View My Scores bug fix #270
Any record can be found by its primary key in the database if you
know what type of object you are looking for. Any time you have
an id passed in as a parameter, you can use the top level class
name which corresponds to the primary key in the database.

In this instance, the code was looking for an AssignmentQuestionnaire
object, and the id was available, so using the code
self.assignment_questionnaire.find_by... throws a NoMethodError
because the 'find' method is a class method of each model, not
an instance method called on a specific object.