Bug fix for #270
Someone renamed the assignment_questionnaires in the Questionnaire
object from assignment_questionnaire to assignment_quesetionnaires.

I renamed this instance variable appropriately in the
compute_weighted_score method.

Finding the assignment questionnaire with AssignmentQuestionnaires.find
seemed like a good solution before, but by using the Questionnaire's
assignment_quesetionnaires foreign key relationship, the total
AssignmentQuestionnaires are actually filtered in scope. This seems
to prevent AssignmentQuestionnaires with the same Assignment id
from being found in this method. Only AssignmentQuestionnaires that
belong to this Questionnaire are searched.