E 912
This project aims at improving score views and performing testing for
score models and views. This is basically an improvement to the E805
project submitted in the OSS assignment. The basic requirement of the
project is to optimize the views for displaying scores in a
comprehensive manner and to test the methods for fetching scores and the
views to display them.