Added assignment form object controller I have added an assignment form object controller, but I'm not sure if this is necessary. I left it for now, but I started editing the assignments/new view file instead. I created another subfolder in views/assignments called new that basically functions like the assignments/edit folder. It will hold partial views for sign up topics, due dates, etc.
Due dates and sign up topics are strange and don't lend well to a form object, since they require an assignment form id. This can be done in the form of a transcation, where we save the assignment, get its id and use it for the other objects, but it doesn't seem like the best solution.