Tested for teamed assignments.
After patching the code to prevent the "Delete Selected Hyperlinks" button from
showing up when not the submisstion owner, the button no longer shows. However,
teamed assignments still displays the other team members submissions when listing
submissions to review. This is because the app/views/submitted_content/_main form partial
does not check if the @current_user is in the same team.

Modified _main.html.erb to pass only what is needed to the _hyperlink.html.erb form partial
(specifically, the assignment_participant). The passed in participant into _main.html.erb
was not consistently the same object.

In the _hyperlink form partial, the @current_user.name is checked against the
assignment_participant.handle. If they are the same, then the button shows up, otherwise it
is not rendered. Here, the assignment_participant is whoever submitted the content that
will be displayed by the current _hyperlink form partial rendering. This check allows
the _hyperlink form partial to still be used when looking at the current user's work
to delete submitted content.