Pushing the fix for 392.
_main.html.erb extracts the AssignmentParticipant that is passed in
from participant. This may be an array (when reviewing other people's work),
it may not be (when looking at your own work), but usually it is an array.
From there, the assignment_participant (value extracted from the
passed in participant) is passed to _hyperlinks.html.erb to be checked.
In _hyperlinks, if the @current_user is the owner of the submission,
then the button renders.