Create new function get_emails_signed_up_users
This change is the first part of refactoring the mail_signed_up_users
function.  The mail_signed_up_users can be described as basically
four cases based on if there are subtopics and if it is a team
assignment:

  1. If the assignment has no subtopics and if is not a team
     assignment, then send an email to all assignment participants

  2. If the assignment has no subtopics and is a team assignment,
     then send a mail to all team members

  3. If the assignment has subtopics and is not a team assignment,
     then send a mail to all signed up users for each of the
     subtopics

  4. If the assignment has subtopics and is a team assignment, then
     send a mail to all team members for each of the subtopics

This first in a series of commits introduces a function called
get_emails_signed_up_users which will get the email addresses for
cases 3 and 4 above.

This commit also removes the logic to find the emails from the
mail_signed_up_users.