Continue to refactor mail_signed_up_users
This commit is the second part of refactoring this function.
In the first part (commit id da2a18f440da197e0d6a02866d4442001ea2d0c9)
We pulled out the logic to get emails in 2 of 4 cases handled in
this function.  In this commit we address cases 1 and 2.  In these
two cases, there are no subtopics.