E1788. OSS project Maroon: Heatmap fixes
Assigned a unique id to all tables based on the round number and included a class "scoresTable" for each table and initialised sorting features based on class.

Added only true values in the resource object returned for the TA, ensured that only courses in which he/she is TA of will be sent back to the view.