Copied small amount of code from old pull request #1270.
Modified method name from update_is_waitlisted to remove_team_from_waitlist.

Team seems to be correctly removed from waitlist when a member of that team lists the signup sheet. However, it isn't until the next time the signup sheet is accessed that the change is shown in the view.