Add assignment_or_course_is_available?()
This new method will replace update_is_available() and update_is_available_2(). This optimization reduces redundant checks on the TA. It should not matter how the TA relates to the instructor. We should only check if the TA has privileges to view this course. If they do then all other checks are not needed.