Add pagination option for all users and RSpec tests for cooresponding Rspec Tests
Added a fourth option to display all users on a single page. Modified the per_page parameter to dynamically set the number of users per page based on user input or default values. Added tests for all pagination options to verify the functionality and behavior of the paginate_list method.